Replace sample with a hidden from_iter implementation for double-ended-stream
