+++ src/stream/double_ended/mod.rs
@@ -3,12 +3,14 @@ mod nth_back;
 mod rfind;
 mod rfold;
 mod try_rfold;
+mod from_iter;
 
 use next_back::NextBackFuture;
 use nth_back::NthBackFuture;
 use rfind::RFindFuture;
 use rfold::RFoldFuture;
 use try_rfold::TryRFoldFuture;
+pub use from_iter::{from_iter, FromIter};
 
 extension_trait! {
     use crate::stream::Stream;
@@ -113,10 +115,9 @@ extension_trait! {
             ```
             # fn main() { async_std::task::block_on(async {
             #
-            use async_std::stream::Sample;
-            use async_std::stream::double_ended::DoubleEndedStreamExt;
+            use async_std::stream::double_ended::{self, DoubleEndedStreamExt};
 
-            let mut s = Sample::from(vec![7u8]);
+            let mut s = double_ended::from_iter(vec![7u8]);
 
             assert_eq!(s.next_back().await, Some(7));
             assert_eq!(s.next_back().await, None);
@@ -141,10 +142,9 @@ extension_trait! {
             ```
             # fn main() { async_std::task::block_on(async {
             #
-            use async_std::stream::Sample;
-            use async_std::stream::double_ended::DoubleEndedStreamExt;
+            use async_std::stream::double_ended::{self, DoubleEndedStreamExt};
 
-            let mut s = Sample::from(vec![1u8, 2, 3, 4, 5]);
+            let mut s = double_ended::from_iter(vec![1u8, 2, 3, 4, 5]);
 
             let second = s.nth_back(1).await;
             assert_eq!(second, Some(4));
@@ -172,10 +172,9 @@ extension_trait! {
             ```
             # fn main() { async_std::task::block_on(async {
             #
-            use async_std::stream::Sample;
-            use async_std::stream::double_ended::DoubleEndedStreamExt;
+            use async_std::stream::double_ended::{self, DoubleEndedStreamExt};
 
-            let mut s = Sample::from(vec![1u8, 2, 3, 4, 5]);
+            let mut s = double_ended::from_iter(vec![1u8, 2, 3, 4, 5]);
 
             let second = s.rfind(|v| v % 2 == 0).await;
             assert_eq!(second, Some(4));
@@ -202,10 +201,9 @@ extension_trait! {
             ```
             # fn main() { async_std::task::block_on(async {
             #
-            use async_std::stream::Sample;
-            use async_std::stream::double_ended::DoubleEndedStreamExt;
+            use async_std::stream::double_ended::{self, DoubleEndedStreamExt};
 
-            let s = Sample::from(vec![1, 2, 3, 4, 5]);
+            let s = double_ended::from_iter(vec![1u8, 2, 3, 4, 5]);
 
             let second = s.rfold(0, |acc, v| v + acc).await;
 
@@ -237,10 +235,9 @@ extension_trait! {
             ```
             # fn main() { async_std::task::block_on(async {
             #
-            use async_std::stream::Sample;
-            use async_std::stream::double_ended::DoubleEndedStreamExt;
+            use async_std::stream::double_ended::{self, DoubleEndedStreamExt};
 
-            let s = Sample::from(vec![1, 2, 3, 4, 5]);
+            let s = double_ended::from_iter(vec![1u8, 2, 3, 4, 5]);
             let sum = s.try_rfold(0, |acc, v| {
                 if (acc+v) % 2 == 1 {
                     Ok(v+3)
